Split app.js phase 2c: extract modal-manager and time-inputs modules
Shared time-picker/schedule-mode state moved to state.js. In-app tests
pass (132/132).
