+++ src/theme.js
@@ -5,9 +5,9 @@ import { tauriAPI } from './tauri-api.js';
 import { getCurrentWebview } from '@tauri-apps/api/webview';
 import { resolveMicrosoftStorePackage, isVersionHigher, getLatestVersionPlatformKey } from './update-banner.js';
 import { updateOverrideAllButtonVisibility, refreshUninstallButtonState } from './settings.js';
+import { stopHelperUiRefreshLoop } from './modal-manager.js';
 import {
     saveData,
-    stopHelperUiRefreshLoop,
     setLanguagePickerOpen,
     applySettingsLanguage,
     applyFormattedCurrentVersion,
